requirements.txt
pandas
numpy
nltk
sklearn
sparse_dot_topn
vowpalwabbit
PyYaml

